// automated-test/utils/interfaces.ts
export interface ReportData {
    tcType: string;
    reportCategoryId: string;
    fullname: string;
    gender: string;
    province_id: string;
    tipe_korban: string;
    description: string;
    lat: string;
    lng: string;
    address_location: string;
    urgency: string;
    userfilePath: string;
  }